Extract pure logic from the audio-cutter dialog into tested modules
Move the pure formatting, segment/range math, word-mark estimation, and
speech-detection helpers out of the 4,120-line audio-cutter-dialog.tsx into
three tested modules under src/lib/editor/audio/:

- format.ts: clamp, formatSeconds, getFileBaseName, getErrorMessage
- segments.ts: Segment/TimeRange types + sort/normalize/keep-range math,
  plus the pure detectSpeechSegmentsFromAnalysis / getSegmentSkipRangesFromAnalysis
- word_marks.ts: getApproximateWordMarks, getApproximateWordPlaybackRange,
  applyWordMarkTimeOverrides, getActiveWordMarkIndex

The dialog changes are deletions plus import lines only — no behavior change.
DOM/React/wavesurfer-coupled helpers and the WeakMap-cached segmentation stay
in the component. Adds 48 characterization tests (node:test). Dialog shrinks
from 4,120 to 3,613 lines.

Implements plans/009-split-audio-cutter-dialog.md.
