+++ src/lib/editor/audio/format.ts
@@ -1,0 +1,23 @@
+// Pure formatting/number utilities extracted from the audio-cutter dialog.
+// No DOM, React, or wavesurfer coupling — safe to unit test in isolation.
+
+export function clamp(value: number, min: number, max: number) {
+  return Math.min(max, Math.max(min, value));
+}
+
+export function formatSeconds(value: number) {
+  const totalMs = Math.max(0, Math.round(value * 1000));
+  const minutes = Math.floor(totalMs / 60_000);
+  const seconds = Math.floor((totalMs % 60_000) / 1000);
+  const milliseconds = totalMs % 1000;
+  return `${String(minutes).padStart(2, "0")}:${String(seconds).padStart(2, "0")}.${String(milliseconds).padStart(3, "0")}`;
+}
+
+export function getFileBaseName(filename: string) {
+  const dotIndex = filename.lastIndexOf(".");
+  return dotIndex > 0 ? filename.slice(0, dotIndex) : filename;
+}
+
+export function getErrorMessage(error: unknown, fallback: string) {
+  return error instanceof Error && error.message ? error.message : fallback;
+}
